Add Player.cs - важное
Вынесены настройки ИИ в отдельный класс, добавлен клас игрока, AIController больше не монобех, UniTask добавлена

+++ Assets/Scripts/Services/Interfaces/ICameraController.cs
@@ -4,7 +4,7 @@ namespace Services
 {
     public interface ICameraController : IService
     {
-        float MoveDuration { get; }
+        int MoveDuration { get; }  // время в милисек
         Camera MainCamera { get; }
     }
 }
